Fix extralight search issue
It was turning on the incorrect extralight with some vehicles. Because it was considering the root bone as part of the turret hierarchy, so if there is a extralight bone that is a direct child of the root it would be considered part of the turret.
For example, with following hierchies:
- Turret:       chassis > turret_1base > weapon_1a
- Extralight: chassis > extralight_1